requirements.txt
# Core runtime
requests==2.32.3
pillow==10.4.0
numpy<2.0.0,>=1.23.0

# AI/ML
torch>=2.2.0
transformers==4.44.2
ultralytics==8.2.100
piq==0.8.0
opencv-python==4.10.0.84

# Utils
tqdm==4.66.5

# Backend
Django==5.0.7
djangorestframework==3.15.2

# Frontend
streamlit==1.37.1
pandas==2.2.2
plotly==5.24.1

# Optional extras
scipy==1.14.1
python-dotenv==1.0.1
instagrapi==2.1.1
yt-dlp==2024.8.6
